+++ common/steamcraft/common/block/BlockCastIronLamp.java
@@ -363,9 +363,9 @@ public class BlockCastIronLamp extends BlockContainer
 		} else
 		{
 			int l = world.getBlockMetadata(i, j, k);
-			double d = (double)((float)i + 0.5F) + (double)(random.nextFloat() - 0.5F) * 0.20000000000000001D;
-			double d1 = (double)((float)j + 1.0F) + (double)(random.nextFloat() - 0.5F) * 0.20000000000000001D;
-			double d2 = (double)((float)k + 0.5F) + (double)(random.nextFloat() - 0.5F) * 0.20000000000000001D;
+			double d = i + 0.5F + (random.nextFloat() - 0.5F) * 0.20000000000000001D;
+			double d1 = j + 1.0F + (random.nextFloat() - 0.5F) * 0.20000000000000001D;
+			double d2 = k + 0.5F + (random.nextFloat() - 0.5F) * 0.20000000000000001D;
 			double d3 = 0.2199999988079071D;
 			double d4 = 0.27000001072883606D;
